Restructure main method
The main method is now just calling a 'run' function that would either
return an error or nil.

If the function returns nil, the main function will call os.Exit(0) and
end execution. If an error is returned, the main function will print
that error to stderr and then exit with os.Exit(1).