Fix bug regarding endpoint selection
Due to the way that the flags package does set the flags value the
variable 'endpoint' was always either an empty string or the actual
endpoint url.

The original code was checking if the variable is not nil. That was never
the case. This meant that when the user did not set a endpoint the empty
string was used as endpoint. This caused issues.