Remove Default Value of Region Flag
If no "--region" flag was set, the region was set to "us-east-1". This
happens before the new DynamoDB session is created. When the session is
created it will try to find the credentials (e.g. read from
~/.aws/credentials). This file might contain a region which should be
used.

That means, the region flag of ddbt should only be used to overwrite the
default defined elsewhere, but never set a region by default itself.