Make Table Name a Normal Argument
To allow simpler usage of the tool the table is no longer a flag but an
argument. The reason for this change is to make the tool easier to use.
The name of the table is the only required input of the tool. It is the
main focus of the whole tool: truncating THIS table.

Therefore, requiring a flag for it is not appropriate.

Before the change:

    ddbt --table-name TestTable

After the change:

    ddbt TestTable

Maybe later both will possible to align the arguments of ddbt better
with the arguments of the AWS cli. But for now let's keep it simple.