Minor Improvements
This commit contains a few improvements to the overall user experience
of the command. This includes the following:

  1. Add --help flag to show help text
  2. Add --version flag to output version of ddbt
  3. Update the usage message (better formatting, new flags)

Furthermore, the code was a bit re-arranged to allow simpler development
of the help and version flags. This means moving the code to parse the
command line flags out of the 'newConfig()' method and to a point in the
flow where it is nearly the first thing that is done.

Further re-arrangements of this sort are to be expected in future
commits.