Create Single Context and Use Everywhere
This change is meant to improve the programs behaviour when the user
interrupts the program. For example by pressing CTRL+C.

There has been a context before, that was used with the AWS SDK clients,
but this context was not bound to the OS signals.

With this patch this changes. Although I am not a 100% sure if the
current implementation is final.

The code has been taken from Mat Ryers blog:

https://medium.com/@matryer/make-ctrl-c-cancel-the-context-context-bd006a8ad6ff

Closes #24